+++ src/config/index.js
@@ -264,9 +264,19 @@ async function getConfig() {
 				},
 				tags: [
 					{
+						name: "DOC",
+						description:
+							"Endpoints used for the conversion of DOC documents",
+					},
+					{
 						name: "DOCX",
 						description:
 							"Endpoints used for the conversion of DOCX documents",
+					},
+					{
+						name: "HTML",
+						description:
+							"Endpoints used for the conversion of HTML documents",
 					},
 					{
 						name: "PDF",