+++ src/plugins/docx-to-html/plugin.test.js
@@ -49,6 +49,10 @@ describe("DOCX-to-HTML conversion plugin", () => {
 		const dom = new JSDOM(body);
 
 		expect(isHtml(body)).toBe(true);
+		// String found in header of the test document
+		expect(dom.window.document.querySelector("header").textContent).toMatch(
+			"I am a header"
+		);
 		// String found in first heading of the test document
 		expect(dom.window.document.querySelector("h1").textContent).toBe(
 			"Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nunc ac faucibus odio. "
@@ -60,6 +64,10 @@ describe("DOCX-to-HTML conversion plugin", () => {
 			].textContent
 		).toMatch(
 			/Nullam venenatis commodo imperdiet. Morbi velit neque, semper quis lorem quis, efficitur dignissim ipsum. Ut ac lorem sed turpis imperdiet eleifend sit amet id sapien$/
+		);
+		// String found in footer of the test document
+		expect(dom.window.document.querySelector("footer").textContent).toMatch(
+			"I am a footer"
 		);
 		// Expect all images to be embedded
 		dom.window.document.querySelectorAll("img").forEach((image) => {